+++ app/src/main/java/io/github/rikuyu/contactlensreminder/data/repository/LensSettingRepositoryImpl.kt
@@ -14,4 +14,8 @@ class LensSettingRepositoryImpl @Inject constructor(
     }
 
     override fun getAllLensSetting(): LensSettingValue = localDataSource.getAllLensSetting()
+
+    override fun createChannel() {
+        localDataSource.createChannel()
+    }
 }
