+++ app/src/main/java/io/github/rikuyu/contactlensreminder/domain/repository/LensSettingRepository.kt
@@ -7,4 +7,6 @@ interface LensSettingRepository {
     fun saveAllLensSetting(lensSettingValue: LensSettingValue)
 
     fun getAllLensSetting(): LensSettingValue
+
+    fun createChannel()
 }
