+++ app/src/main/java/io/github/rikuyu/contactlensreminder/domain/repository/ReminderRepository.kt
@@ -17,6 +17,4 @@ interface ReminderRepository {
     fun getIsDarkTheme(): Boolean
 
     fun switchIsDarkTheme()
-
-    fun createChannel()
 }
