+++ app/src/main/java/io/github/rikuyu/contactlensreminder/domain/usecase/reminder/ReminderUseCase.kt
@@ -17,6 +17,4 @@ data class ReminderUseCase @Inject constructor(
     val getIsDarkTheme: GetIsDarkTheme,
 
     val switchIsDarkTheme: SwitchIsDarkTheme,
-
-    val createChannel: CreateChannel,
 )
