+++ app/src/main/java/io/github/rikuyu/contactlensreminder/ui/screens/lens_setting/components/LensSettingScreen.kt
@@ -18,7 +18,6 @@ import androidx.compose.ui.graphics.Color
 import androidx.compose.ui.platform.LocalContext
 import androidx.compose.ui.res.stringResource
 import androidx.compose.ui.unit.dp
-import androidx.core.app.NotificationManagerCompat
 import androidx.hilt.navigation.compose.hiltViewModel
 import androidx.navigation.NavController
 import io.github.rikuyu.contactlensreminder.R
@@ -53,8 +52,6 @@ fun LensSettingScreen(
     }
 
     var dialogState by remember { mutableStateOf(false) }
-
-    val notificationManager = NotificationManagerCompat.from(context)
 
     Scaffold(topBar = {
         TopAppBar(
