+++ app/src/main/res/values-en/strings.xml
@@ -16,12 +16,6 @@
     <string name="review_launch_complete">launchReviewFlow(a, r).addOnCompleteListener_en</string>
     <string name="review_launch_failure">launchReviewFlow(a, r).addOnFailureListener_en</string>
 
-    <!-- SCHEDULE_EXACT_ALARM -->
-    <string name="exact_alarm_dialog_title">Request to change settings</string>
-    <string name="exact_alarm_dialog_message">To use this application, please allow "Alarms and Reminders" from the Settings screen (please turn the setting ON)</string>
-    <string name="exact_alarm_dialog_positive_button_label">Open Settings</string>
-    <string name="exact_alarm_dialog_event_label">alarmManager.canScheduleExactAlarms() == %s</string>
-
     <!-- TopScreen -->
     <string name="remain">left</string>
     <string name="day">days</string>
