+++ app/src/test/java/io/github/rikuyu/contactlensreminder/ui/screens/top/ReminderViewModelTest.kt
@@ -36,9 +36,6 @@ class ReminderViewModelTest {
     @MockK
     lateinit var switchIsDarkTheme: SwitchIsDarkTheme
 
-    @MockK
-    lateinit var createChannel: CreateChannel
-
     @Before
     fun setup() {
         MockKAnnotations.init(this)
@@ -49,8 +46,7 @@ class ReminderViewModelTest {
             resetReminder,
             getIsShowOnBoarding,
             getIsDarkTheme,
-            switchIsDarkTheme,
-            createChannel
+            switchIsDarkTheme
         )
     }
 
@@ -73,7 +69,6 @@ class ReminderViewModelTest {
         every { getReminderSetting.invoke() } returns defaultReminderValue
         every { saveReminderSetting.invoke(any()) } returns Unit
         every { startReminder.invoke() } returns Unit
-        every { createChannel.invoke() } returns Unit
         every { getIsShowOnBoarding.invoke() } returns true
         every { getIsDarkTheme.invoke() } returns false
 
@@ -92,7 +87,6 @@ class ReminderViewModelTest {
             viewModel.onEvent(event)
             saveReminderSetting.invoke(any())
             startReminder.invoke()
-            createChannel.invoke()
         }
     }
 }
